# Fix collectstatic sometimes missing updated files.