Django's `collectstatic` command uses file modification time to determine
if it should overwrite a file.  Unfortunately inside VirtualBox shared
folder file modification times are often inconsistent, and this can lead
to newly updated files not being collected, especially when running
`grunt watch`.

Adding `-c` to the command tells Django to always overwrite every file,
which fixes the problem, though it does add a slight amount of overhead.