@jwalgran 2015-09-17
Could you quantify "slight?" The edit/build/test loop is already longer than I would like it to be. I am not sure if we have a choice, since we need a collectstatic to do the right thing, but maybe trying to attack the problem through VirtualBox (NFS shares?) could help avoid the performance penalty.